+++ database/migrations/2025_08_04_142004_create_purchase_request_items_table.php
@@ -8,7 +8,8 @@ return new class extends Migration {
   /**
    * Run the migrations.
    */
-  public function up(): void {
+  public function up(): void
+  {
     Schema::create('purchase_request_items', function (Blueprint $table) {
       $table->ulid("id")->primary();
       $table->foreignUlid('purchase_request_id')->references('id')->on('purchase_requests')->cascadeOnDelete();
@@ -17,9 +18,9 @@ return new class extends Migration {
       $table->string('item_name')->nullable();
       $table->timestamp('required_date')->nullable();
       $table->foreignUlid('target_warehouse_id')->nullable()->references('id')->on('warehouses')->nullOnDelete();
-      $table->unsignedInteger('quantity')->default(1);
-      $table->unsignedInteger('ordered_quantity')->default(0);
-      $table->unsignedInteger('remaining_quantity')->storedAs('quantity - ordered_quantity');
+      $table->double('quantity')->default(1);
+      $table->double('ordered_quantity')->default(0);
+      $table->double('remaining_quantity')->storedAs('quantity - ordered_quantity');
       $table->foreignUlid('unit_id')->nullable()->references('id')->on('units')->nullOnDelete();
       $table->string("unit_name")->nullable();
       $table->double('conversion_factor')->nullable()->default(1);
@@ -32,7 +33,8 @@ return new class extends Migration {
   /**
    * Reverse the migrations.
    */
-  public function down(): void {
+  public function down(): void
+  {
     Schema::dropIfExists('purchase_request_items');
   }
 };
